+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as WeekView } from './WeekView';
+export { default as MonthView } from './MonthView';
+export { default as EventFormView } from './EventFormView';
